implement project mock handlers
